requirements.txt
django==5.1.1
djangorestframework==3.15.2
ruff==0.6.4
Faker==28.4.1
factory-boy==3.3.1